refactor(backend): API
* update backend docs
    * this is not complete because I expect further changes to be made and these go into A LOT of detail
* move some non-arma 3 endpoints into the generic endpoint
